Create increasing_array_cses.cpp
You are given an array of n integers. You want to modify the array so that it is increasing, i.e., every element is at least as large as the previous element.

On each move, you may increase the value of any element by one. We need to find the  minimum number of moves required.